Move all startup files to /plugin
- Benefits: lua files show up when checking startuptime

- Cons: can get hectic if more startup files are added
