+++ dot_config/nvim/init.lua
@@ -1,22 +1,8 @@
--- require'utils.modules'.source(false)
-
 local function disable_default_plugins()
   -- FIXME: WSL register bug
   vim.g.loaded_clipboard_provider = 1
-  vim.g.loaded_gzip = 1
-  vim.g.loaded_tar = 1
-  vim.g.loaded_tarPlugin = 1
-  vim.g.loaded_zip = 1
-  vim.g.loaded_zipPlugin = 1
-  vim.g.loaded_getscript = 1
-  vim.g.loaded_getscriptPlugin = 1
-  vim.g.loaded_vimball = 1
-  vim.g.loaded_vimballPlugin = 1
   vim.g.loaded_matchit = 1
   vim.g.loaded_matchparen = 1
-  vim.g.loaded_2html_plugin = 1
-  vim.g.loaded_logiPat = 1
-  vim.g.loaded_rrhelper = 1
   vim.g.loaded_netrw = 1
   vim.g.loaded_netrwPlugin = 1
   vim.g.loaded_netrwSettings = 1
@@ -25,14 +11,4 @@ end
 
 disable_default_plugins()
 
-require("utils.keymaps")
-
-require("utils.globals")
-
-require("settings")
-
-vim.cmd("source " .. vim.fn.stdpath("config") .. "/autocommands.vim")
-
-require("plugins")
-
-require("mappings")
+-- Loading moved to top level plugins folder
