+++ dot_config/nvim/plugin/20-globals.lua
