+++ dot_config/nvim/plugin/30-settings.lua.tmpl
