+++ dot_config/nvim/plugin/40-autocommands.vim
